Separate what a multiplexer is from how it is run (#547)
Adding the build commit and the installation-derived namespace pushed
`multiplexer.rs` past the 1000-line policy ceiling. The fix is to split
it rather than to raise the limit, and the seam was already visible.

Version and identity parsing is the one part of multiplexer handling
that is pure: it never resolves an executable, never spawns a child,
never reads the environment. Everything around it does all three. That
difference is worth a module boundary on its own, and it also means the
parser can be reasoned about and exercised without any of the process
machinery the rest of the file exists to drive.

The types stay re-exported from `multiplexer.rs`, so callers continue
to see a single multiplexer surface and no import churn spreads through
the tree. This is a move, not a rewrite: the parsing logic is byte for
byte what it was.
